requirements.txt
fastapi
python-multipart
uvicorn
asyncio
openai
pydantic
python-dotenv
PyPDF2
pytesseract